requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-2.9.6
python-dotenv-1.0.0